Refactor: kanban, calendar 페이지 /페이지로 통합 처리 #57
