+++ src/routes.tsx
@@ -1,16 +1,14 @@
 import React from 'react';
 
 const SignIn = React.lazy(() => import('./pages/signIn/SignIn'));
-const Calendar = React.lazy(() => import('./pages/main/calendar/Calendar'));
-const Kanban = React.lazy(() => import('./pages/main/kanban/Kanban'));
 const Page404 = React.lazy(() => import('./pages/page404/Page404'));
 const Page500 = React.lazy(() => import('./pages/page500/Page500'));
 const DesiredPosition = React.lazy(() => import('./pages/signUp/DesiredPosition'));
 const Application = React.lazy(() => import('./pages/applications/index'));
+const Home = React.lazy(() => import('./pages/main/index'));
 
 export const privateRoutes = [
-  { path: '/calendar', element: <Calendar /> },
-  { path: '/kanban', element: <Kanban /> },
+  { path: '/', element: <Home /> },
   { path: '/application/*', element: <Application /> },
 ];
